+++ src/main/kotlin/com/codescene/jetbrains/services/api/CodeDeltaService.kt
@@ -43,8 +43,18 @@ class CodeDeltaService(private val project: Project) : CodeSceneService() {
 
     override fun getActiveApiCalls() = CodeSceneCodeVisionProvider.activeDeltaApiCalls
 
+    /**
+     * Performs delta analysis by comparing the current editor content against a baseline.
+     *
+     * The baseline for delta analysis is determined as follows:
+     * - If available, it uses the code/content from the branch creation commit.
+     * - If the branch creation commit is not found or the analysis is run in a detached HEAD state,
+     *   it falls back to comparing against the best score of 10.0.
+     */
     private suspend fun performDeltaAnalysis(editor: Editor) {
-        val oldCode = GitService.getInstance(project).getHeadCommit(editor.virtualFile)
+        val oldCode = GitService
+            .getInstance(project)
+            .getBranchCreationCommitCode(editor.virtualFile)
 
         val delta = getDeltaResponse(editor, oldCode)
 
@@ -56,6 +66,7 @@ class CodeDeltaService(private val project: Project) : CodeSceneService() {
 
         val oldReview = ReviewParams(path, oldCode)
         val newReview = ReviewParams(path, editor.document.text)
+
         val delta = runWithClassLoaderChange { ExtensionAPI.delta(oldReview, newReview) }
 
         if (delta?.oldScore == null) delta?.oldScore = 10.0